requirements.txt
pytest==7.4.0
requests==2.31.0
jsonschema==4.19.0
allure-pytest==2.9.45
selene>=2.0.0b4
webdriver-manager==3.8.6
curlify==2.2.1